requirements.txt
# requirements.txt
Flask
nltk
numpy
tensorflow-cpu
requests # Good to have, though not used in this core version
